+++ src/components/client_type/DeleteClientTypes.tsx
@@ -3,11 +3,11 @@ import { OverlayEventDetail } from '@ionic/core/components';
 import React from 'react';
 import { AxiosError } from 'axios';
 import { process_error_hint } from '../../utils/process_erros_hints';
-import { RefetchFunction } from 'axios-hooks'
-import { AuthProps } from '../../interface/props/auth';
 import API from '../../utils/server';
 import ClientType from '../../interface/client_type';
-import { useAppSelector } from '../../redux/store';
+import { useAppDispatch, useAppSelector } from '../../redux/store';
+import presentNoAuthAlert from '../../utils/present_no_auth_alert';
+import { fetch } from '../../redux/client_types';
 
 export function DeleteClientTypesModal(
   {selected_client_types, onDismiss}: {
@@ -46,12 +46,12 @@ export function DeleteClientTypesModal(
 }
 
 export interface DeleteClientTypesModalControllerProps {
-  refetch_client_types: RefetchFunction<any, any>,
   selected_client_types: Array<ClientType>,
 }
 
 export const DeleteClientTypesModalController: React.FC<DeleteClientTypesModalControllerProps> = (props) => {
   const auth = useAppSelector(state => state.auth);
+  const dispatch = useAppDispatch();
   
   const [present, dismiss] = useIonModal(DeleteClientTypesModal, {
     selected_client_types: props.selected_client_types,
@@ -63,6 +63,10 @@ export const DeleteClientTypesModalController: React.FC<DeleteClientTypesModalCo
     present({
       onWillDismiss: (ev: CustomEvent<OverlayEventDetail>) => {
         if (ev.detail.role === 'confirm') {
+          if (auth === null) {
+            return presentNoAuthAlert(presentAlert);
+          }
+          
           Promise.allSettled(ev.detail.data.map(async (role: ClientType) => {
             await API
               .delete_with_auth(auth!, `client_type?id=eq.${role.id}`)
@@ -70,7 +74,7 @@ export const DeleteClientTypesModalController: React.FC<DeleteClientTypesModalCo
           .then((results) => {
             for (const result of results) {
               if (result.status === "rejected" && result.reason instanceof AxiosError) {
-                props.refetch_client_types();
+                dispatch(fetch(auth));
                 presentAlert({
                   header: "Ошибка",
                   subHeader: result.reason.response?.statusText,
@@ -80,7 +84,7 @@ export const DeleteClientTypesModalController: React.FC<DeleteClientTypesModalCo
                 return;
               }
             }
-            props.refetch_client_types();
+            dispatch(fetch(auth));
             presentAlert({
               header: "Типы клиентов удалены",
               buttons: ["Ок"]
